+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes for each version of this project will be documented in this 
 
 ## 9.1.0
 
+### General
+- `IgxGrid`, `IgxTreeGrid`, `IgxHierarchicalGrid`
+    - **Behavioral Change** - When a column is sortable sort indicator is always visible. The column is sorted when click on it.
+
 ### Themes
 - **Breaking Change**  Change the default `$legacy-support` value to false in the `igx-theme` function.
 
@@ -17,6 +21,17 @@ All notable changes for each version of this project will be documented in this 
     ```typescript
     public pinningConfiguration: IPinningConfig = { columns: ColumnPinningPosition.End };
     ```
+    - Added functionality for column selection.
+    - `selected` property has been added to the IgxColumnComponent; Allows you to set whether the column is selected.
+    - `selectable` property has been added to the IgxColumnComponent; Allows you to set whether the column is selectable.
+    - `onColumnSelectionChange` event is added for the `IgxGrid`. It is emitted when the column selection is changed.
+    - `excelStyleSelectingTemplate` property is introduced to IgxGrid, which allows you to set a custom template for the selecting a column in the Excel Style Filter.
+    - `selectedColumns` API method is added for the `IgxGrid`. It allows to get all selected columns.
+    - `selectColumns` API method is added for the `IgxGrid`. It allows to select columns by passing array of IgxColumnComponent or column fields.
+    - `deselectColumns` API method is added for the `IgxGrid`. It allows to deselect columns by passing array of IgxColumnComponent or column fields.
+    - `deselectAllColumns` API method is added for the `IgxGrid`. It allows to deselect all columns.
+    - `getSelectedColumnsData` API method is added for the `IgxGrid`. It allows to get the selected columns data.
+
 - `IgxCombo`:
     - Added `autoFocusSearch` input that allows to manipulate the combo's opening behavior. When the property is `true` (by default), the combo's search input is focused on open. When set to `false`, the focus goes to the combo items container, which can be used to prevent the software keyboard from activating on mobile devices when opening the combo.
 
@@ -24,7 +39,7 @@ All notable changes for each version of this project will be documented in this 
 - `igxSlider` have full right-to-left (RTL) support.
 
 ## 9.0.1
-- **Breaking Changes** 
+- **Breaking Changes**
     - Remove `$base-color` from igx-typography. The igx-typography class now inherits the parent color.
 
 ## 9.0.0
@@ -108,10 +123,6 @@ All notable changes for each version of this project will be documented in this 
     - `IgxGridExcelStyleFilteringComponent` and `IgxAdvancedFilteringDialogComponent` can now be hosted outside of the grid in order to provide the same experience as the built-in filtering UI.
     - `expandRow(rowID)`/`collapseRow(rowID)`/`toggleRow(rowID)` API methods are added for the `igxHierarchicalGrid`. They allow expanding/collapsing a row by its id.
     - `onRowToggle` event is added for the `igxHierarchicalGrid`. It is emitted when the expanded state of a row is changed.
-    - `IgxOverlayService`:
-        - `setOffset` method added. It offsets the content along the corresponding axis by the provided amount.
-    - `IgxToggleDirective`:
-        - `setOffset` method added. It offsets the content along the corresponding axis by the provided amount.
     - `IgxRowDragGhost` directive is added. It allows providing a custom template for the drag ghost when dragging a row.
     ```html
     <igx-grid #grid1 [data]="remote | async" primaryKey="ProductID"
@@ -170,6 +181,12 @@ All notable changes for each version of this project will be documented in this 
 
 - `IgxDropDown`:
     - `clearSelection` method is added, which can be used to deselect the selected dropdown item
+
+- `IgxToggleDirective`:
+    - `setOffset` method added. It offsets the content along the corresponding axis by the provided amount.
+
+- `IgxOverlayService`:
+    - `setOffset` method added. It offsets the content along the corresponding axis by the provided amount.
 
 - `IgxCircularProgressBar`:
     - added `IgxProgressBarGradientDirective` to allow providing custom circular progress SVG gradients. Providing a custom gradient via a template is as easy as writing: